avoid gcc warning

+++ include/header1.c
@@ -459,7 +459,7 @@ static void exit_continuation (void)
   dump_object ((object *) result, 0);
   fprintf (stdout, "\n");
   fprintf (stderr, "{total ticks: %" PRIu64 " gc ticks: %" PRIu64 "}\n", program_end_time - program_start_time, gc_ticks);
-  exit((int)result);
+  exit((int)(intptr_t)result);
 }
 
 static void toplevel (void);
